requirements.txt
streamlit==1.28.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.3
plotly==5.18.0
Pillow==10.1.0
barcode==0.15.1
python-barcode==0.15.1
qrcode==7.4.2
streamlit-authenticator==0.2.3
bcrypt==4.1.1
pyzbar==0.1.9
opencv-python==4.8.1.78
streamlit-camera-input-live==0.2.0
